cinema-app/requirements.txt
flask ~= 1.1
flask_wtf ~= 1.2
grpcio-tools ~= 1.30
Jinja2 ~= 2.11
pytest ~= 5.4
markupsafe == 2.0.1